Implemented map functionality
Shows a list of last 10 trips where the user can see graphically in the map and the details in a dialog.
Changed userId from Travel to be mandatory.
Fixed an error with culture formats.
Fixed an issue with Realm DB: Primary keys were not auto generated nor auto incremented.
Fixed an error with google api response.
Fixed an error in web api where if the user doesn't have any travels and asks for stats, an unhandled exception was thrown.
Removed the WiFi check before update the scanned results.