Updated realm
Changed intensity network detection values.
Fixed an error with general statistics: If the DB was empyu, an exception was thrown.
Fixed an error with resources.
Fixed an issue with list adapter in MainActivity: Concurrency synchronization problem.
Fixed an error with MainActivity intent: Travel ID is a long, not an integer anymore.
Renamed RestClient and Scanner classes.